Slim Boogu pipeline: inline scheduler adapter, drop LoRA mixin and rewriter
Reduce the boogu pipeline package from 7 files to 4 by removing dead and
misplaced code, keeping the default T2I/TI2I inference path unchanged.

- Inline set_flow_match_timesteps into pipeline_boogu.py (single caller) and
  delete flow_match_boogu.py, per the "inline single-caller helpers" rule.
- Replace the image_processor.preprocess override (which duplicated the parent
  VaeImageProcessor wholesale) with a thin override that only derives the
  Boogu max_pixels/max_side_length target size, then delegates to the parent.
  Verified bit-identical output across sizes/constraints (max diff 0.0).
- Remove BooguImageLoraLoaderMixin / lora_pipeline.py: LoRA is unused on the
  inference path, and the mixin belongs in loaders/ by diffusers convention.
- Remove the instruction-rewriter feature entirely (static_skills.py,
  instruct_reasoner_static_skills.py, and ~1100 lines of rewriter methods,
  state, and public kwargs). It was gated by use_rewrite_text_instruction
  (default False) and unused by every example/test; the skills files were its
  only consumers.

Net: -2255 / +74 lines. End-to-end TI2I inference reproduces the standalone
reference (mean pixel diff 8.8, unchanged from before), and the boogu test
suite shows the same pre-existing baseline (11 failed / 19 passed / 4 skipped,
the 11 being unrelated MLLM device-placement failures). check_copies and
check_dummies pass.
